+++ Changes.md
@@ -14,6 +14,7 @@ API
 - EventLoop : Added `BlockedUIThreadExecution` context manager.
 - ScriptNode : Added support for cancellation of execution and serialisation.
 - ValuePlug : Improved warning emitted if cached value has unexpected type.
+- GafferUI.FileMenu :  `__open()` is now public as `openScript()`
 
 0.59.4.0 (relative to 0.59.3.0)
 ========